requirements.txt
discord.py>=2.3.2
python-dotenv>=1.0.0
requests>=2.31.0
apscheduler>=3.10.4
beautifulsoup4>=4.12.0
lxml>=4.9.3
aiohttp>=3.9.0
feedparser>=6.0.10
